+++ src/constants/index.ts
@@ -9,10 +9,22 @@ export const RENDERED_PAGE_COMPONENT_RADIUS = 6;
 export const RENDERED_PAGE_LAUNCH_COMPONENT_ROWS = 4;
 export const MAXIMUM_FILE_SIZE = 3; //In MB
 export const PASSWORD_REGEX = /^(?=.*\d)(?=.*[a-z]).{8,}$/gm;
-export const EMAIL_REGEX = /^.+@.+\..+$/gm;
+export const EMAIL_REGEX = new RegExp("^.+@.+..+$", "gmi");
+export const MAILTO_REGEX = new RegExp("^(mailto:).+@.+..+", "gmi");
+export const TEL_REGEX = new RegExp("^(tel:)[0-9]+", "gmi");
 export const TOKEN_AUTH_ERROR = "auth/invalid-token";
 export const MIN_WEBSITE_WIDTH = "275px";
 export const FOOTER_HEIGHT = "100px";
+export const GDPR_CONSENT_STORAGE_KEY = "gdprConsent";
+export const LINK_VALIDATOR_REGEX = new RegExp(
+  "^(https?:\\/\\/)?" + // protocol
+    "((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.)+[a-z]{2,}|" + // domain name
+    "((\\d{1,3}\\.){3}\\d{1,3}))" + // OR ip (v4) address
+    "(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*" + // port and path
+    "(\\?[;&a-z\\d%_.~+=-]*)?" + // query string
+    "(\\#[-a-z\\d_]*)?$",
+  "i"
+);
 
 export const COMPONENT_ANIMATIONS = [
   "bounce",