+++ src/modules/components/site-content/full-width/index.tsx
@@ -6,6 +6,7 @@ import { Root } from "./style";
 const FullWidthContent = ({ children, ...rest }: any) => {
   useEffect(() => {
     document.body.style.backgroundColor = GLOBAL_LIGHT_BG;
+    window.scrollTo(0, 0);
   }, []);
 
   return (