+++ src/routes/paths.ts
@@ -5,12 +5,13 @@ const routes = {
   pages: "/dashboard/pages",
   page: "/dashboard/page",
   profile: "/dashboard/profile",
-  support: "/public/support",
+  support: "/support",
   faq: "/public/faq",
   notFound: "/public/404",
   about: "/public/about",
   privacy: "/public/privacy",
   terms: "/public/terms",
+  cookies: "/public/cookies",
 };
 
 export default routes;