+++ src/utils/validators/url.ts
@@ -1,13 +1,16 @@
+import { LINK_VALIDATOR_REGEX, MAILTO_REGEX, TEL_REGEX } from "../../constants";
+
 export const isUrlValid = (str: string | undefined) => {
-  if (!str || str.length < 1) return false;
-  var pattern = new RegExp(
-    "^(https?:\\/\\/)?" + // protocol
-      "((([a-z\\d]([a-z\\d-]*[a-z\\d])*)\\.)+[a-z]{2,}|" + // domain name
-      "((\\d{1,3}\\.){3}\\d{1,3}))" + // OR ip (v4) address
-      "(\\:\\d+)?(\\/[-a-z\\d%_.~+]*)*" + // port and path
-      "(\\?[;&a-z\\d%_.~+=-]*)?" + // query string
-      "(\\#[-a-z\\d_]*)?$",
-    "i"
-  ); // fragment locator
-  return !!pattern.test(str);
+  if (!str || !str.length) return false;
+  return str.match(LINK_VALIDATOR_REGEX);
+};
+
+export const isMailtoUrlValid = (str: string | undefined) => {
+  if (!str || !str.length) return false;
+  return str.match(MAILTO_REGEX);
+};
+
+export const isTelUrlValid = (str: string | undefined) => {
+  if (!str || !str.length) return false;
+  return str.match(TEL_REGEX);
 };